+++ Lab6/printercards.h
@@ -11,19 +11,19 @@ class PrinterCards
 private:
     std::vector<std::string> namefile_cards;        //масив імен файлів картинок карт
 
-    int width;         //ширина карти
-    int height;        //висота карти
+    float width;         //ширина карти
+    float height;        //висота карти
 
-    int size_mark;                // розмір мітки
-    int x_mark_move_players;      //координата x малювання мітки біля гравця чий хід
-    int y_mark_move_player1;      //координата y малювання мітки біля гравця1
-    int y_mark_move_player2;      //координата y малювання мітки біля гравця2
+    float size_mark;                // розмір мітки
+    float x_mark_move_players;      //координата x малювання мітки біля гравця чий хід
+    float y_mark_move_player1;      //координата y малювання мітки біля гравця1
+    float y_mark_move_player2;      //координата y малювання мітки біля гравця2
 
-    int x_cards_deck;       //координата x малювання карти колоди на gameboard
-    int y_cards_deck;       //координата y розміщення карти колоди на gameboard
+    float x_cards_deck;       //координата x малювання карти колоди на gameboard
+    float y_cards_deck;       //координата y розміщення карти колоди на gameboard
 
-    int x_top_card;         //координата x малювання останньої покладеної карти колоди на gameboard
-    int y_top_card;         //координата y малювання останньої покладеної карти колоди на gameboard
+    float x_top_card;         //координата x малювання останньої покладеної карти колоди на gameboard
+    float y_top_card;         //координата y малювання останньої покладеної карти колоди на gameboard
 
     QGraphicsItem* ptr_card_deck = nullptr;               //вказівник на малюнок колоди
     QGraphicsItem* ptr_card_converted_card = nullptr;     //вказівник на малюнок карти яку вибрали (8-ою або джокером)
@@ -33,18 +33,18 @@ public:
     PrinterCards();
 
             // встановлення значень координат малювання колоди і верхньої карти
-    void set_coordinate(int x_deck, int y_deck, int x_top_c, int y_top_c);
+    void set_coordinate(float x_deck, float y_deck, float x_top_c, float y_top_c);
             //встановлення розмірів карти   (в пікселях)
-    void set_side(int width, int height);
+    void set_side(float width, float height);
 
-    int get_width();    //отримання ширини карти(в пікселях)
-    int get_height();   //отримання висоти карти (в пікселях)
+    float get_width();    //отримання ширини карти(в пікселях)
+    float get_height();   //отримання висоти карти (в пікселях)
 
-    int get_x_cards_deck();     //отримання координати x малювання колоди (в пікселях)
-    int get_y_cards_deck();     //отримання координати y малювання колоди (в пікселях)
+    float get_x_cards_deck();     //отримання координати x малювання колоди (в пікселях)
+    float get_y_cards_deck();     //отримання координати y малювання колоди (в пікселях)
 
-    int get_x_top_card();       //отримання координати x малювання верхньої карти (в пікселях)
-    int get_y_top_card();       //отримання координати y малювання верхньої карти (в пікселях)
+    float get_x_top_card();       //отримання координати x малювання верхньої карти (в пікселях)
+    float get_y_top_card();       //отримання координати y малювання верхньої карти (в пікселях)
 
             // малювання рубашки карти на коордах x y розміру width height
     QGraphicsItem* print_back(QGraphicsScene*& scene, int x, int y);
@@ -70,7 +70,7 @@ public:
     void erase_converted_card();
 
             //малювання мітки відображення ходу (чий хід)
-    void print_marc_move(QGraphicsScene*& scene, int current_player, int x_mark_players, int y_mark_player1, int y_mark_player2, int size);
+    void print_marc_move(QGraphicsScene*& scene, int current_player, float x_mark_players, float y_mark_player1, float y_mark_player2, float size);
             //зміна положення мітки відображення ходу (чий хід) (зміна ходу)
     void print_change_move(QGraphicsScene*& scene, int current_player);
             //видалення мітки відображення ходу (чий хід)
